Comments and cleanup.
Add TODO comments in a few places where RSTudio is issuing warnings.

Add some new lines and adjusted style (re-apply rstyle after applying a newline before the first argument of lapply() statements) at the bottom of the file, for readability.

Add comments to a few fairly long inline functions that seem to be defined in multiple places with only minor variations in definition.  It looks like the code could be considerably shortened and simplified by extracting these to top-level functions with a few extra arguments.